% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/gee_methods.R,
%   R/glm_methods.R, R/parfrailty_methods.R, R/standardize_custom.R
\name{print.std_coxph}
\alias{print.std_coxph}
\alias{print.summary_std_coxph}
\alias{print.std_gee}
\alias{print.summary_std_gee}
\alias{print.std_glm}
\alias{print.summary.parfrailty}
\alias{print.std_parfrailty}
\alias{print.std_helper}
\title{Prints summary of Cox regression standardization fit}
\usage{
\method{print}{std_coxph}(x, ...)

\method{print}{summary_std_coxph}(x, ...)

\method{print}{std_gee}(x, ...)

\method{print}{summary_std_gee}(x, ...)

\method{print}{std_glm}(x, ...)

\method{print}{summary.parfrailty}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{std_parfrailty}(x, ...)

\method{print}{std_helper}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.parfrailty"}.}

\item{\dots}{not used.}

\item{digits}{the number of significant digits to use when printing.}
}
\description{
This is a \code{print} method for class \code{"summary.std_coxph"}.

This is a \code{print} method for class \code{"summary_std_gee"}.

This is a \code{print} method for class \code{"summary.parfrailty"}.
}
\examples{


## See documentation for standardize_coxph


## See documentation for standardize_gee


## See documentation for frailty

}
\seealso{
\code{\link{standardize_coxph}}

\code{\link{standardize_gee}}

\code{\link{parfrailty}}
}
\author{
Arvid Sjolander

Arvid Sjolander and Elisabeth Dahlqwist
}
