% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfrailty_methods.R
\name{print.summary.parfrailty}
\alias{print.summary.parfrailty}
\title{Print method for parametric frailty fits}
\usage{
\method{print}{summary.parfrailty}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class "parfrailty"}

\item{digits}{Number of digits to print}

\item{...}{Not used}
}
\value{
The object being printed, invisibly
}
\description{
Print method for parametric frailty fits
}
