% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sandwich}
\alias{sandwich}
\title{Compute the sandwich variance components from a model fit}
\usage{
sandwich(fit, data, weights, t, fit.detail)
}
\arguments{
\item{fit}{A fitted model object of class glm, coxph, ah, or survfit}

\item{data}{The data used to fit the model}

\item{weights}{Optional weights}

\item{t}{Optional fixed time point for survival objects}

\item{fit.detail}{Additional information for survival objects, see Details}
}
\value{
A list consisting of the Fisher information matrix (I) and the Score equations (U)
}
\description{
Compute the sandwich variance components from a model fit
}
