% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_custom.R
\name{standardize}
\alias{standardize}
\title{Get standardized estimates using the g-formula with a custom model}
\usage{
standardize(
  arguments,
  data,
  fitter,
  predict_fun,
  values,
  B = NULL,
  ci_level = 0.95,
  contrasts = NULL,
  references = NULL,
  seed = NULL,
  times = NULL,
  transforms = NULL
)
}
\arguments{
\item{arguments}{The arguments to be used in the fitter function as a \code{list}.}

\item{data}{The data.}

\item{fitter}{The function to call to fit the data.}

\item{predict_fun}{The function used to predict the means/probabilities
for a new data set on the response level. For survival data,
this should be a matrix where each column is the time, and each
row the data.}

\item{values}{A named list or data.frame specifying the variables and values
at which marginal means of the outcome will be estimated.}

\item{B}{Number of nonparametric bootstrap resamples. Default is \code{NULL} (no bootstrap).}

\item{ci_level}{Coverage probability of confidence intervals.}

\item{contrasts}{A vector of contrasts in the following format:
If set to \code{"difference"} or \code{"ratio"}, then \eqn{\psi(x)-\psi(x_0)}
or \eqn{\psi(x) / \psi(x_0)} are constructed, where \eqn{x_0} is a reference
level specified by the \code{reference} argument. Has to be be \code{NULL}
if no references are specified.}

\item{references}{A vector of references in the following format:
If \code{contrasts} is not \code{NULL}, the desired reference level(s).}

\item{seed}{The seed to use with the nonparametric bootstrap.}

\item{times}{For use with survival data. Set to \code{NULL} otherwise.}

\item{transforms}{A vector of transforms in the following format:
If set to \code{"log"}, \code{"logit"}, or \code{"odds"}, the standardized
mean \eqn{\theta(x)} is transformed into \eqn{\psi(x)=log\{\theta(x)\}},
\eqn{\psi(x)=log[\theta(x)/\{1-\theta(x)\}]}, or
\eqn{\psi(x)=\theta(x)/\{1-\theta(x)\}}, respectively.
If the vector is \code{NULL}, then \eqn{\psi(x)=\theta(x)}.}
}
\value{
An object of class \code{std_custom}.
This is basically a list with components estimates and fit for the outcome model.
}
\description{
Get standardized estimates using the g-formula with a custom model
}
\details{
Let \eqn{Y}, \eqn{X}, and \eqn{Z} be the outcome, the exposure, and a
vector of covariates, respectively.
\code{standardize} uses a
model to estimate the standardized
mean \eqn{\theta(x)=E\{E(Y|X=x,Z)\}},
where \eqn{x} is a specific value of \eqn{X},
and the outer expectation is over the marginal distribution of \eqn{Z}.
With survival data, \eqn{Y=I(T > t)},
and a vector of different time points \code{times} (\eqn{t}) can be given,
where \eqn{T} is the uncensored survival time.
}
\examples{

set.seed(6)
n <- 100
Z <- rnorm(n)
X <- rnorm(n, mean = Z)
Y <- rbinom(n, 1, prob = (1 + exp(X + Z))^(-1))
dd <- data.frame(Z, X, Y)
prob_predict.glm <- function(...) predict.glm(..., type = "response")

x <- standardize(
  arguments = list(
    formula = Y ~ X * Z,
    family = "binomial"
  ),
  predict_fun = prob_predict.glm,
  fitter = "glm",
  data = dd,
  values = list(X = seq(-1, 1, 0.1)),
  B = 100,
  references = 0,
  contrasts = "difference"
)
x
plot(x)
plot(x, reference = 0, contrast = "difference")
plot(x, reference = 0, contrast = "difference", plot_ci = FALSE)

require(survival)
prob_predict.coxph <- function(object, newdata, times){
  fit.detail <- suppressWarnings(basehaz(object))
  cum.haz <- fit.detail$hazard[sapply(times, function(x) max(which(fit.detail$time <= x)))]
  predX <- predict(object = object, newdata = newdata, type = "risk")
  res <- matrix(NA, ncol = length(times), nrow = length(predX))
  for (ti in seq_len(length(times))){
    res[, ti] <- exp(-predX*cum.haz[ti])
  }
  res
}
set.seed(68)
n <- 500
Z <- rnorm(n)
X <- rnorm(n, mean = Z)
T <- rexp(n, rate = exp(X + Z + X * Z)) # survival time
C <- rexp(n, rate = exp(X + Z + X * Z)) # censoring time
U <- pmin(T, C) # time at risk
D <- as.numeric(T < C) # event indicator
dd <- data.frame(Z, X, U, D)
x <- standardize(
  arguments = list(
    formula = Surv(U, D) ~ X + Z + X * Z,
    method = "breslow",
    x = TRUE,
    y = TRUE
  ),
  fitter = "coxph",
  data = dd,
  times = 1:5,
  predict_fun = prob_predict.coxph,
  values = list(X = c(-1, 0, 1)),
  B = 100,
  references = 0,
  contrasts = "difference"
)
x
plot(x)
plot(x, reference = 0, contrast = "difference")
plot(x, reference = 0, contrast = "difference", plot_ci = FALSE)
}
\references{
Rothman K.J., Greenland S., Lash T.L. (2008).
\emph{Modern Epidemiology}, 3rd edition.
Lippincott, Williams \& Wilkins.

Sjolander A. (2016).
Regression standardization with the R-package stdReg.
\emph{European Journal of Epidemiology} \bold{31}(6), 563-574.

Sjolander A. (2016).
Estimation of causal effect measures with the R-package stdReg.
\emph{European Journal of Epidemiology} \bold{33}(9), 847-858.
}
