% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/gee_methods.R,
%   R/parfrailty_methods.R
\name{summary.std_coxph}
\alias{summary.std_coxph}
\alias{summary.std_gee}
\alias{summary.parfrailty}
\alias{summary.std_parfrailty}
\title{Summarizes Cox regression standardization fit}
\usage{
\method{summary}{std_coxph}(
  object,
  times,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{summary}{std_gee}(
  object,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{summary}{parfrailty}(
  object,
  ci_type = "plain",
  ci_level = 0.95,
  digits = max(3L, getOption("digits") - 3L),
  ...
)

\method{summary}{std_parfrailty}(
  object,
  times,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"std_parfrailty"}.}

\item{times}{numeric, indicating the times at which to summarize. It defaults to
the specified value(s) of the argument \code{t} in the \code{stdCox}
function.}

\item{ci_type}{string, indicating the type of confidence intervals. Either
"plain", which gives untransformed intervals, or "log", which gives
log-transformed intervals.}

\item{ci_level}{desired coverage probability of confidence intervals, on
decimal form.}

\item{transform}{a string. If set to \code{"log"}, \code{"logit"}, or
\code{"odds"}, the standardized survival function \eqn{\theta(t,x)} is
transformed into \eqn{\psi(t,x)=log\{\theta(t,x)\}},
\eqn{\psi(t,x)=log[\theta(t,x)/\{1-\theta(t,x)\}]}, or
\eqn{\psi(t,x)=\theta(t,x)/\{1-\theta(t,x)\}}, respectively. If left
unspecified, \eqn{\psi(t,x)=\theta(t,x)}.}

\item{contrast}{a string. If set to \code{"difference"} or \code{"ratio"},
then \eqn{\psi(t,x)-\psi(t,x_0)} or \eqn{\psi(t,x) / \psi(t,x_0)} are
constructed, where \eqn{x_0} is a reference level specified by the
\code{reference} argument.}

\item{reference}{must be specified if \code{contrast} is specified.}

\item{\dots}{not used.}

\item{digits}{the number of significant digits to use when printing..}
}
\description{
This is a \code{summary} method for class \code{"std_coxph"}.

This is a \code{summary} method for class \code{"std_gee"}.

This is a \code{summary} method for class \code{"parfrailty"}.

This is a \code{summary} method for class \code{"std_parfrailty"}.
}
\examples{

## See documentation for standardize_coxph


## See documentation for standardize_gee

## See documentation for parfrailty


## See documentation for standardize_parfrailty

}
\seealso{
\code{\link{standardize_coxph}}

\code{\link{standardize_gee}}

\code{\link{parfrailty}}

\code{\link{standardize_parfrailty}}
}
\author{
Arvid Sjolander

Arvid Sjolander and Elisabeth Dahlqwist.
}
