% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/gee_methods.R,
%   R/glm_methods.R, R/parfrailty_methods.R, R/standardize_custom.R
\name{plot.std_coxph}
\alias{plot.std_coxph}
\alias{plot.std_gee}
\alias{plot.std_glm}
\alias{plot.std_parfrailty}
\alias{plot.std_helper}
\title{Plots Cox regression standardization fit}
\usage{
\method{plot}{std_coxph}(
  x,
  plot_ci = TRUE,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  legendpos = "bottomleft",
  ...
)

\method{plot}{std_gee}(
  x,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{plot}{std_glm}(
  x,
  plot_ci = TRUE,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{plot}{std_parfrailty}(
  x,
  plot_ci = TRUE,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  legendpos = "bottomleft",
  ...
)

\method{plot}{std_helper}(
  x,
  plot_ci = TRUE,
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"std_parfrailty"}.}

\item{plot_ci}{logical, indicating whether confidence intervals should be
added to the plot.}

\item{ci_type}{string, indicating the type of confidence intervals. Either
"plain", which gives untransformed intervals, or "log", which gives
log-transformed intervals.}

\item{ci_level}{desired coverage probability of confidence intervals, on
decimal form.}

\item{transform}{a string. If set to \code{"log"}, \code{"logit"}, or
\code{"odds"}, the standardized survival function \eqn{\theta(t,x)} is
transformed into \eqn{\psi(t,x)=log\{\theta(t,x)\}},
\eqn{\psi(t,x)=log[\theta(t,x)/\{1-\theta(t,x)\}]}, or
\eqn{\psi(t,x)=\theta(t,x)/\{1-\theta(t,x)\}}, respectively. If left
unspecified, \eqn{\psi(t,x)=\theta(t,x)}.}

\item{contrast}{a string. If set to \code{"difference"} or \code{"ratio"},
then \eqn{\psi(t,x)-\psi(t,x_0)} or \eqn{\psi(t,x) / \psi(t,x_0)} are
constructed, where \eqn{x_0} is a reference level specified by the
\code{reference} argument.}

\item{reference}{must be specified if \code{contrast} is specified.}

\item{legendpos}{position of the legend; see help for \code{legend}.}

\item{\dots}{further arguments passed on to plot.default.}
}
\description{
This is a \code{plot} method for class \code{"std_coxph"}.

This is a \code{plot} method for class \code{"std_gee"}.

This is a \code{plot} method for class \code{"std_glm"}.

This is a \code{plot} method for class \code{"std_parfrailty"}.
}
\examples{

## See documentation for standardize_coxph


## See documentation for standardize_gee

# see standardize_glm



## See documentation for standardize_parfrailty

}
\seealso{
\code{\link{standardize_coxph}}

\code{\link{standardize_gee}}

\code{\link{standardize_parfrailty}}
}
\author{
Arvid Sjolander
}
