% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/gee_methods.R,
%   R/glm_methods.R, R/parfrailty_methods.R
\name{plot.stdCoxph}
\alias{plot.stdCoxph}
\alias{plot.stdGee}
\alias{plot.stdGLM}
\alias{plot.stdParfrailty}
\title{Plots Cox regression standardization fit}
\usage{
\method{plot}{stdCoxph}(
  x,
  plot.CI = TRUE,
  CI.type = "plain",
  CI.level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  legendpos = "bottomleft",
  ...
)

\method{plot}{stdGee}(
  x,
  CI.type = "plain",
  CI.level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{plot}{stdGLM}(
  x,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{plot}{stdParfrailty}(
  x,
  plot.CI = TRUE,
  CI.type = "plain",
  CI.level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  legendpos = "bottomleft",
  ...
)
}
\arguments{
\item{x}{an object of class \code{"stdParfrailty"}.}

\item{plot.CI}{logical, indicating whether confidence intervals should be
added to the plot.}

\item{CI.type}{string, indicating the type of confidence intervals. Either
"plain", which gives untransformed intervals, or "log", which gives
log-transformed intervals.}

\item{CI.level}{desired coverage probability of confidence intervals, on
decimal form.}

\item{transform}{a string. If set to \code{"log"}, \code{"logit"}, or
\code{"odds"}, the standardized survival function \eqn{\theta(t,x)} is
transformed into \eqn{\psi(t,x)=log\{\theta(t,x)\}},
\eqn{\psi(t,x)=log[\theta(t,x)/\{1-\theta(t,x)\}]}, or
\eqn{\psi(t,x)=\theta(t,x)/\{1-\theta(t,x)\}}, respectively. If left
unspecified, \eqn{\psi(t,x)=\theta(t,x)}.}

\item{contrast}{a string. If set to \code{"difference"} or \code{"ratio"},
then \eqn{\psi(t,x)-\psi(t,x_0)} or \eqn{\psi(t,x) / \psi(t,x_0)} are
constructed, where \eqn{x_0} is a reference level specified by the
\code{reference} argument.}

\item{reference}{must be specified if \code{contrast} is specified.}

\item{legendpos}{position of the legend; see help for \code{legend}.}

\item{\dots}{further arguments passed on to plot.default.}

\item{ci_type}{A string, indicating the type of confidence intervals. Either "plain", which
gives untransformed intervals, or "log", which gives log-transformed intervals.}

\item{ci_level}{Coverage probability of confidence intervals.}
}
\description{
This is a \code{plot} method for class \code{"stdCoxph"}.

This is a \code{plot} method for class \code{"stdGee"}.

This is a \code{plot} method for class \code{"stdGLM"}.

This is a \code{plot} method for class \code{"stdParfrailty"}.
}
\examples{

## See documentation for stdCoxph


## See documentation for stdGee

# see standardize_glm



## See documentation for stdParfrailty

}
\seealso{
\code{\link{stdCoxph}}

\code{\link{stdGee}}

\code{\link{stdParfrailty}}
}
\author{
Arvid Sjolander
}
