% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{print.summary.stdGLM}
\alias{print.summary.stdGLM}
\alias{print.stdGLM}
\title{Prints summary of GLM regression standardization fit}
\usage{
\method{print}{summary.stdGLM}(x, ...)

\method{print}{stdGLM}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.stdGLM"}.}

\item{...}{unused}
}
\description{
Prints summary of GLM regression standardization fit
}
