% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/gee_methods.R,
%   R/glm_methods.R, R/parfrailty_methods.R, R/standardize_custom.R
\name{print.stdCoxph}
\alias{print.stdCoxph}
\alias{print.summary.stdCoxph}
\alias{print.stdGee}
\alias{print.summary.stdGee}
\alias{print.std_glm}
\alias{print.summary.parfrailty}
\alias{print.stdParfrailty}
\alias{print.summary.stdParfrailty}
\alias{print.std_helper}
\title{Prints summary of Cox regression standardization fit}
\usage{
\method{print}{stdCoxph}(x, ...)

\method{print}{summary.stdCoxph}(x, ...)

\method{print}{stdGee}(x, ...)

\method{print}{summary.stdGee}(x, ...)

\method{print}{std_glm}(x, ...)

\method{print}{summary.parfrailty}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{stdParfrailty}(x, ...)

\method{print}{summary.stdParfrailty}(x, ...)

\method{print}{std_helper}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.stdParfrailty"}.}

\item{\dots}{not used.}

\item{digits}{the number of significant digits to use when printing.}
}
\description{
This is a \code{print} method for class \code{"summary.stdCoxph"}.

This is a \code{print} method for class \code{"summary.stdGee"}.

This is a \code{print} method for class \code{"summary.parfrailty"}.

This is a \code{print} method for class \code{"summary.stdParfrailty"}.
}
\examples{


## See documentation for stdCoxph


## See documentation for stdGee


## See documentation for frailty



## See documentation for stdParfrailty

}
\seealso{
\code{\link{stdCoxph}}

\code{\link{stdGee}}

\code{\link{parfrailty}}

\code{\link{stdParfrailty}}
}
\author{
Arvid Sjolander

Arvid Sjolander and Elisabeth Dahlqwist
}
