% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/glm_methods.R,
%   R/standardize_custom.R
\name{print.std_surv}
\alias{print.std_surv}
\alias{print.std_glm}
\alias{print.std_custom}
\title{Prints summary of regression standardization fit}
\usage{
\method{print}{std_surv}(x, ...)

\method{print}{std_glm}(x, ...)

\method{print}{std_custom}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"std_glm"}, \code{"std_surv"} or \code{"std_custom"}.}

\item{\dots}{unused}
}
\value{
The object being printed, invisibly.
}
\description{
Prints summary of regression standardization fit
}
