% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sandwich}
\alias{sandwich}
\title{Compute the sandwich variance components from a model fit}
\usage{
sandwich(fit, data, weights, t, fit.detail)
}
\arguments{
\item{fit}{A fitted model object of class \link[stats]{glm}, \link[survival]{coxph}, ah, or \link[survival]{survfit}}

\item{data}{The data used to fit the model}

\item{weights}{Optional weights}

\item{t}{Optional fixed time point for survival objects}

\item{fit.detail}{For Cox models, the result of running \link[survival]{coxph.detail} on the model fit}
}
\value{
A list consisting of the Fisher information matrix (I) and the Score equations (U)
}
\description{
Compute the sandwich variance components from a model fit
}
