% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{standardize_glm}
\alias{standardize_glm}
\title{Get regression standardized estimates from a glm}
\usage{
standardize_glm(
  formula,
  data,
  family = gaussian,
  values,
  se = TRUE,
  clusterid,
  case.control = FALSE,
  p.population,
  matched.density.cases,
  matched.density.controls,
  matching.variable
)
}
\arguments{
\item{values}{A named list or data.frame specifying the variables and values
at which marginal means of the outcome will be estimated.}

\item{se}{If \code{TRUE}, calculate standard errors and the asymptotic covariance matrix.}

\item{fit}{The fitted object as returned by glm.}
}
\value{
An object of class std. This is basically a list with components estimates and covariance
}
\description{
Get regression standardized estimates from a glm
}
\examples{

n <- 1000
Z <- rnorm(n)
X <- rnorm(n, mean=Z)
Y <- rbinom(n, 1, prob=(1+exp(X+Z))^(-1))
dd <- data.frame(Z, X, Y)
standardize_glm(formula=Y~X+Z+X*Z,family="binomial", data=dd, values = list(X = 0:1))
}
