% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdReg2-package.R
\docType{package}
\name{stdReg2-package}
\alias{stdReg2}
\alias{stdReg2-package}
\title{stdReg2: Regression Standardization for Causal Inference}
\description{
Contains more modern tools for causal inference using regression standardization. Four general classes of models are allowed; generalized linear models, conditional generalized estimating equation models, Cox proportional hazards models and shared frailty gamma-Weibull models. Sjolander, A. (2016) \doi{10.1007/s10654-016-0157-3}. Also includes functionality for doubly robust estimation for these model classes in some special cases.
}
\author{
\strong{Maintainer}: Michael C Sachs \email{sachsmc@gmail.com}

Authors:
\itemize{
  \item Arvid Sjölander
  \item Erin E Gabriel
  \item Johan Sebastian Ohlendorff
}

}
