% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_methods.R, R/gee_methods.R,
%   R/glm_methods.R, R/parfrailty_methods.R
\name{summary.stdCoxph}
\alias{summary.stdCoxph}
\alias{summary.stdGee}
\alias{summary.stdGLM}
\alias{summary.parfrailty}
\alias{summary.stdParfrailty}
\title{Summarizes Cox regression standardization fit}
\usage{
\method{summary}{stdCoxph}(
  object,
  t,
  CI.type = "plain",
  CI.level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{summary}{stdGee}(
  object,
  CI.type = "plain",
  CI.level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{summary}{stdGLM}(
  object,
  ci_type = "plain",
  ci_level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)

\method{summary}{parfrailty}(
  object,
  CI.type = "plain",
  CI.level = 0.95,
  digits = max(3L, getOption("digits") - 3L),
  ...
)

\method{summary}{stdParfrailty}(
  object,
  t,
  CI.type = "plain",
  CI.level = 0.95,
  transform = NULL,
  contrast = NULL,
  reference = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"stdParfrailty"}.}

\item{t}{numeric, indicating the times at which to summarize. It defaults to
the specified value(s) of the argument \code{t} in the \code{stdCox}
function.}

\item{CI.type}{string, indicating the type of confidence intervals. Either
"plain", which gives untransformed intervals, or "log", which gives
log-transformed intervals.}

\item{CI.level}{desired coverage probability of confidence intervals, on
decimal form.}

\item{transform}{a string. If set to \code{"log"}, \code{"logit"}, or
\code{"odds"}, the standardized survival function \eqn{\theta(t,x)} is
transformed into \eqn{\psi(t,x)=log\{\theta(t,x)\}},
\eqn{\psi(t,x)=log[\theta(t,x)/\{1-\theta(t,x)\}]}, or
\eqn{\psi(t,x)=\theta(t,x)/\{1-\theta(t,x)\}}, respectively. If left
unspecified, \eqn{\psi(t,x)=\theta(t,x)}.}

\item{contrast}{a string. If set to \code{"difference"} or \code{"ratio"},
then \eqn{\psi(t,x)-\psi(t,x_0)} or \eqn{\psi(t,x) / \psi(t,x_0)} are
constructed, where \eqn{x_0} is a reference level specified by the
\code{reference} argument.}

\item{reference}{must be specified if \code{contrast} is specified.}

\item{\dots}{not used.}

\item{ci_type}{A string, indicating the type of confidence intervals. Either "plain", which
gives untransformed intervals, or "log", which gives log-transformed intervals.}

\item{ci_level}{Coverage probability of confidence intervals.}

\item{digits}{the number of significant digits to use when printing..}
}
\description{
This is a \code{summary} method for class \code{"stdCoxph"}.

This is a \code{summary} method for class \code{"stdGee"}.

This is a \code{summary} method for class \code{"stdGLM"}.

This is a \code{summary} method for class \code{"parfrailty"}.

This is a \code{summary} method for class \code{"stdParfrailty"}.
}
\examples{

## See documentation for stdCoxph


## See documentation for stdGee

# see standardize_glm
## See documentation for parfrailty


## See documentation for stdParfrailty

}
\seealso{
\code{\link{stdCoxph}}

\code{\link{stdGee}}

\code{\link{parfrailty}}

\code{\link{stdParfrailty}}
}
\author{
Arvid Sjolander

Arvid Sjolander and Elisabeth Dahlqwist.
}
