% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfrailty_methods.R
\name{summary.parfrailty}
\alias{summary.parfrailty}
\title{Summarizes parfrailty fit}
\usage{
\method{summary}{parfrailty}(
  object,
  ci_type = "plain",
  ci_level = 0.95,
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"parfrailty"}.}

\item{ci_type}{string, indicating the type of confidence intervals. Either
"plain", which gives untransformed intervals, or "log", which gives
log-transformed intervals.}

\item{ci_level}{desired coverage probability of confidence intervals, in
decimal form.}

\item{digits}{the number of significant digits to use when printing..}

\item{\dots}{not used.}
}
\value{
An object of class "summary.parfrailty", which is a list that contains relevant summary statistics about the fitted model
}
\description{
This is a \code{summary} method for class \code{"parfrailty"}.
}
\examples{
## See documentation for parfrailty

}
\seealso{
\code{\link{parfrailty}}
}
\author{
Arvid Sjölander and Elisabeth Dahlqwist.
}
